move Button, ButtonSlider, Loader to folder atoms and Footer to folder organisms
