--- src/components/Loading.jsx
@@ -1,5 +1,0 @@
-function Loading() {
-  return null;
-}
-
-export default Loading;
