+++ src/components/atoms/Loader.jsx
@@ -1,0 +1,5 @@
+function Loader() {
+  return null;
+}
+
+export default Loader;
