+++ src/components/molecules/DetailLayananItem.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../Button';
+import Button from '../atoms/Button';
 import PropTypes from 'prop-types';
 
 function DetailLayananItem({ title, image, description, buttonText, invertedPosition, href }) {
